Скролл работает на treeview; интерполяционный алгоритм работает корректно

+++ RC-FE Design - Analysis and synthesis/MathModel/StateOfIndividual.cs
@@ -15,6 +15,6 @@ namespace FractalElementDesigner.MathModel
         /// <summary>
         /// Оценка особи
         /// </summary>
-        public int Rate { get; set; }
+        public double Rate { get; set; }
     }
 }
